+++ libs/react/src/components/application/authentication/protectedRoute.tsx
@@ -6,14 +6,14 @@ import {
   useSearchParams,
 } from 'react-router-dom';
 import { ErrorFallback, SignupPage, Spinner } from '@coldpbc/components';
-import { GlobalSizes } from '@coldpbc/enums';
+import { ErrorType, GlobalSizes } from '@coldpbc/enums';
 import ColdContext from '../../../context/coldContext';
 import { useLDClient } from 'launchdarkly-react-client-sdk';
 import useSWR from 'swr';
 import { axiosFetcher } from '@coldpbc/fetchers';
 import { get, has, isEmpty, isUndefined } from 'lodash';
 import { PolicySignedDataType } from '@coldpbc/interfaces';
-import { useAuth0Wrapper, useOrgSWR } from '@coldpbc/hooks';
+import { useAuth0Wrapper, useColdContext, useOrgSWR } from '@coldpbc/hooks';
 import { SurveyPayloadType } from '@coldpbc/interfaces';
 import { withErrorBoundary } from 'react-error-boundary';
 import { ErrorPage } from '../errors/errorPage';
@@ -29,6 +29,8 @@ const _ProtectedRoute = () => {
     orgId,
   } = useAuth0Wrapper();
   const { auth0Options } = useContext(ColdContext);
+
+  const { logError } = useColdContext();
 
   const ldClient = useLDClient();
 
@@ -110,13 +112,15 @@ const _ProtectedRoute = () => {
       } catch (e) {
         if (has(e, 'error')) {
           if (get(e, 'error') === 'login_required') {
+            logError(e, ErrorType.Auth0Error);
             await loginWithRedirect();
           }
           if (get(e, 'error') === 'consent_required') {
+            logError(e, ErrorType.Auth0Error);
             await loginWithRedirect();
           }
         }
-        console.error(e);
+        logError(e, ErrorType.Auth0Error);
       }
     };
 
@@ -152,8 +156,15 @@ const _ProtectedRoute = () => {
   }
 
   if (error || initialSurveySWR.error || signedPolicySWR.error) {
-    const errorObj = error || initialSurveySWR.error || signedPolicySWR.error;
-    console.error(errorObj);
+    if (error) {
+      logError(error, ErrorType.Auth0Error);
+    }
+    if (initialSurveySWR.error) {
+      logError(initialSurveySWR.error, ErrorType.SWRError);
+    }
+    if (signedPolicySWR.error) {
+      logError(signedPolicySWR.error, ErrorType.SWRError);
+    }
     return <ErrorPage />;
   }
 
@@ -175,7 +186,7 @@ const _ProtectedRoute = () => {
 };
 
 export const ProtectedRoute = withErrorBoundary(_ProtectedRoute, {
-  FallbackComponent: (props) => <ErrorFallback />,
+  FallbackComponent: (props) => <ErrorFallback {...props} />,
   onError: (error, info) => {
     console.error('Error occurred in ProtectedRoute: ', error);
   },