+++ libs/react/src/components/molecules/footprintDetailCard/footprintDetailCard.tsx
@@ -8,6 +8,8 @@ import { useFlags } from 'launchdarkly-react-client-sdk';
 import { withErrorBoundary } from 'react-error-boundary';
 import { ErrorFallback } from '../../application';
 import { useOrgSWR } from '../../../hooks/useOrgSWR';
+import { useColdContext } from '@coldpbc/hooks';
+import { ErrorType } from '@coldpbc/enums';
 
 export interface FootprintDetailCardProps {
   colors: string[];
@@ -24,12 +26,18 @@ function _FootprintDetailCard(
   const [isEmpty, setIsEmpty] = useState(false);
 
   // Get footprint data from SWR
-  const { data } = useOrgSWR<any>(
+  const { data, error } = useOrgSWR<any>(
     ['/categories/company_decarbonization', 'GET'],
     axiosFetcher,
   );
+  const { logError } = useColdContext();
 
   if (isEmpty) {
+    return null;
+  }
+
+  if (error) {
+    logError(error, ErrorType.SWRError);
     return null;
   }
 
@@ -65,7 +73,7 @@ function _FootprintDetailCard(
 }
 
 export const FootprintDetailCard = withErrorBoundary(_FootprintDetailCard, {
-  FallbackComponent: (props) => <ErrorFallback />,
+  FallbackComponent: (props) => <ErrorFallback {...props} />,
   onError: (error, info) => {
     console.error('Error occurred in FootprintDetailCard: ', error);
   },