+++ libs/react/src/components/molecules/journeySpiderChart/journeySpiderChart.tsx
@@ -22,6 +22,8 @@ import { EmptyChart } from './emptyChart';
 import { withErrorBoundary } from 'react-error-boundary';
 import { ErrorFallback } from '../../application/errors/errorFallback';
 import { useOrgSWR } from '../../../hooks/useOrgSWR';
+import { useColdContext } from '@coldpbc/hooks';
+import { ErrorType } from '@coldpbc/enums';
 
 ChartJS.register(
   RadarController,
@@ -47,6 +49,12 @@ function _JourneySpiderChart({ setIsEmptyData }: Props) {
     ['/categories', 'GET'],
     axiosFetcher,
   );
+  const { logError } = useColdContext();
+
+  if (error) {
+    logError(error, ErrorType.SWRError);
+    return null;
+  }
 
   // Update chart data on receiving new data
   const isEmpty =
@@ -164,7 +172,7 @@ function _JourneySpiderChart({ setIsEmptyData }: Props) {
 }
 
 export const JourneySpiderChart = withErrorBoundary(_JourneySpiderChart, {
-  FallbackComponent: (props) => <ErrorFallback />,
+  FallbackComponent: (props) => <ErrorFallback {...props} />,
   onError: (error, info) => {
     console.error('Error occurred in JourneySpiderChart: ', error);
   },