+++ libs/react/src/components/molecules/newsItem/newsItem.tsx
@@ -8,6 +8,18 @@ interface Props {
 export const NewsItem = ({ item }: Props) => {
   const { title, url, image_url, published_at, source_name } = item;
 
+  const getTimeSincePublished = () => {
+    if (!published_at) return null;
+
+    const timeSincePublished = intlFormatDistance(new Date(published_at).valueOf(), new Date());
+
+    return (
+      <span className="mr-4">
+        {timeSincePublished.charAt(0).toUpperCase() + timeSincePublished.slice(1)}
+      </span>
+    );
+  }
+
   return (
     <div className="bg-bgc-accent p-4 w-full text-white flex items-center rounded-lg">
       <div
@@ -19,11 +31,7 @@ export const NewsItem = ({ item }: Props) => {
       <div className="flex-1 flex flex-col">
         <div className="font-bold leading-normal text-md">{title}</div>
         <div className="flex text-xs font-normal mt-2 items-center">
-          {published_at ? (
-            <span className="mr-4">
-              {intlFormatDistance(new Date(published_at).valueOf(), new Date())}
-            </span>
-          ) : null}
+          {getTimeSincePublished()}
           <a
             className="flex items-center bg-gray-50 py-1 px-2 rounded-full"
             target="_blank"