+++ libs/react/src/components/molecules/surveyQuestionContainer/surveyQuestionContainer.tsx
@@ -30,7 +30,7 @@ const _SurveyQuestionContainer = ({
   surveyData,
   setSurveyData,
 }: SurveyQuestionContainerProps) => {
-  const { getOrgSpecificUrl, isLoading, error } = useAuth0Wrapper();
+  const { getOrgSpecificUrl } = useAuth0Wrapper();
   const nextQuestionTransitionClassNames = {
     enter: 'transform translate-y-full',
     enterDone: 'transition ease-out duration-200 transform translate-y-0',
@@ -511,19 +511,6 @@ const _SurveyQuestionContainer = ({
     getActiveQuestion();
   }, [activeKey, surveyData]);
 
-  if (isLoading) {
-    return (
-      <div>
-        <Spinner />
-      </div>
-    );
-  }
-
-  if (error) {
-    console.error(error);
-    return <div></div>;
-  }
-
   if (activeQuestion !== undefined) {
     return (
       <div
@@ -564,7 +551,7 @@ const _SurveyQuestionContainer = ({
 export const SurveyQuestionContainer = withErrorBoundary(
   _SurveyQuestionContainer,
   {
-    FallbackComponent: (props) => <ErrorFallback />,
+    FallbackComponent: (props) => <ErrorFallback {...props} />,
     onError: (error, info) => {
       console.error('Error occurred in SurveyQuestionContainer: ', error);
     },