+++ libs/react/src/components/organisms/surveyRightNav/surveyRightNav.tsx
@@ -55,7 +55,7 @@ const _SurveyRightNav = ({
 };
 
 export const SurveyRightNav = withErrorBoundary(_SurveyRightNav, {
-  FallbackComponent: (props) => <ErrorFallback />,
+  FallbackComponent: (props) => <ErrorFallback {...props} />,
   onError: (error, info) => {
     console.error('Error occurred in SurveyRightNav: ', error);
   },