+++ libs/react/src/components/pages/actionsOverview/actionsOverview.tsx
@@ -58,7 +58,7 @@ const _ActionsOverview = () => {
 };
 
 export const ActionsOverview = withErrorBoundary(_ActionsOverview, {
-  FallbackComponent: (props) => <ErrorFallback />,
+  FallbackComponent: (props) => <ErrorFallback {...props} />,
   onError: (error, info) => {
     console.error('Error occurred in ActionsOverview: ', error);
   },