+++ libs/react/src/enums/errors.ts
@@ -1,0 +1,5 @@
+export enum ErrorType {
+  SWRError = 'SWRError',
+  RenderingError = 'RenderingError',
+  Auth0Error = 'Auth0Error',
+}